+++ qacode/core/loggers/__init__.py
@@ -3,6 +3,7 @@
 
 
 from qacode.core.loggers import logger_manager
+from qacode.core.loggers import logger_messages
 
 
-__all__ = ['logger_manager']
+__all__ = ['logger_manager', 'logger_messages']