+++ qacode/core/webs/controls/control_base.py
@@ -6,6 +6,7 @@ from collections import defaultdict
 from qacode.core.bots.bot_base import BotBase
 from qacode.core.exceptions.control_exception import ControlException
 from qacode.core.exceptions.core_exception import CoreException
+from qacode.core.loggers import logger_messages
 from selenium.common.exceptions import (
     ElementNotVisibleException, NoSuchElementException
 )
@@ -13,47 +14,12 @@ from selenium.webdriver.common.by import By
 from selenium.webdriver.remote.webelement import WebElement
 
 
+MSG = logger_messages
+
+
 class ControlBase(object):
     """Requirements: #35, #70"""
 
-    # Log messages
-    CB_SETTINGS_LOADING = "control | load_settings_keys: loading keys..."
-    CB_SETTINGS_LOADED = "control | load_settings_keys: loaded keys"
-    CB_SEARCH_DISABLED = "control | _load_search: Disabled element search"
-    CB_SEARCH_LOADING = "control | _load_search: searching element..."
-    CB_SEARCH_WAITING = "control | _load_search: waiting for element..."
-    CB_SEARCH_FOUND = "control | _load_search: element found!"
-    CB_SEARCH_FOUND_CHILD = "control | _load_search: element child found!"
-    CB_PROP_DISABLED = ("control | _load_properties: "
-                        "disabled loading for properties")
-    CB_PROP_LOADING = "control | _load_properties: loading properties..."
-    CB_PROP_LOADED = "control | _load_properties: loaded properties"
-    CB_FINDCHILD_LOADING = "control | find_child: searching for selector='{}'"
-    CB_FINDCHILD_LOADED = "control | find_child: child element found"
-    CB_GETTAG_LOADING = "control | get_tag: searching for tag"
-    CB_GETTAG_LOADED = "control | get_tag: found same tag='{}'"
-    CB_TYPETEXT_LOADING = "control | type_text: typing text={}"
-    CB_CLEAR_LOADING = "control | clear: clearing text..."
-    CB_CLEAR_LOADED = "control | clear: cleared text"
-    CB_CLICK_LOADING = "control | click: clicking element..."
-    CB_CLICK_RETRY = "control | click: retry clicking element..."
-    CB_CLICK_LOADED = "control | click: clicked!"
-    CB_GETTEXT_LOADING = "control | get_text: obtaining text..."
-    CB_GETTEXT_FAILED = "control | get_text: failed at obtain text"
-    CB_GETTEXT_LOADED = "control | get_text: text obtained, text='{}'"
-    CB_GETATTRS_LOADING = "control | get_attrs: obtaining attrs..."
-    CB_GETATTRVALUE_LOADING = ("control | get_attr_value: obtaining value "
-                               "for attr_name='{}'")
-    CB_GETATTRVALUE_LOADED = ("control | get_attr_value: obtained"
-                              "attr_name='{}', value='{}'")
-    CB_GETATTRVALUE_FAILED = "control | get_attr_value: fail at obtain value"
-    CB_SETCSSRULE_LOADING = ("control | set_css_value: setting new CSS rule,"
-                             " prop_name={}, prop_value={}")
-    CB_SETCSSRULE_FAILED = "control | set_css_value: failed at set CSS rule"
-    CB_GETCSSRULE_LOADING = "control | get_css_value: obtaining css_value..."
-    CB_GETCSSRULE_LOADED = "control | get_css_value: css_value='{}'"
-    CB_RELOAD_LOADING = "{} | reload: reloading control..."
-    CB_RELOAD_LOADED = "{} | reload: reloaded control"
     # Instance properties
     bot = None
     settings = None
@@ -107,7 +73,7 @@ class ControlBase(object):
 
     def load_settings_keys(self, settings, update=False, default_keys=None):
         """Load default setting for ControlBase instance"""
-        self.bot.log.debug(self.CB_SETTINGS_LOADING)
+        self.bot.log.debug(MSG.CB_SETTINGS_LOADING)
         # generate default dict
         if default_keys is None:
             default_keys = [
@@ -140,30 +106,30 @@ class ControlBase(object):
             setattr(self, key, updated_settings.get(key))
         if update:
             self.settings = updated_settings
-        self.bot.log.debug(self.CB_SETTINGS_LOADED)
+        self.bot.log.debug(MSG.CB_SETTINGS_LOADED)
 
     def _load_search(self, enabled=False, element=None):
         """Load element searching at selenium WebDriver"""
         if not enabled or enabled is None:
-            self.bot.log.debug(self.CB_SEARCH_DISABLED)
+            self.bot.log.debug(MSG.CB_SEARCH_DISABLED)
             return False
-        self.bot.log.debug(self.CB_SEARCH_LOADING)
+        self.bot.log.debug(MSG.CB_SEARCH_LOADING)
         try:
             if element is not None:
                 if not isinstance(element, WebElement):
                     msg = "Child is not instance of WebElement"
                     raise ControlException(msg=msg)
-                self.bot.log.debug(self.CB_SEARCH_FOUND_CHILD)
+                self.bot.log.debug(MSG.CB_SEARCH_FOUND_CHILD)
                 self.element = element
             else:
                 self.element = self.bot.navigation.find_element(
                     self.selector, locator=self.locator)
         except CoreException:
-            self.bot.log.warning(self.CB_SEARCH_WAITING)
+            self.bot.log.warning(MSG.CB_SEARCH_WAITING)
             self.element = self.bot.navigation.find_element_wait(
                 self.selector, locator=self.locator)
         if self.element:
-            self.bot.log.debug(self.CB_SEARCH_FOUND)
+            self.bot.log.debug(MSG.CB_SEARCH_FOUND)
         return True
 
     def _load_properties(self, enabled=False):
@@ -182,9 +148,9 @@ class ControlBase(object):
             self.bot.log.error(msg)
             raise ControlException(msg=msg)
         if not enabled or enabled is None:
-            self.bot.log.debug(self.CB_PROP_DISABLED)
+            self.bot.log.debug(MSG.CB_PROP_DISABLED)
             return False
-        self.bot.log.debug(self.CB_PROP_LOADING)
+        self.bot.log.debug(MSG.CB_PROP_LOADING)
         self.tag = self.get_tag()
         self.text = self.get_text()
         self.is_displayed = self.bot.navigation.ele_is_displayed(self.element)
@@ -192,8 +158,15 @@ class ControlBase(object):
         self.is_selected = self.bot.navigation.ele_is_selected(self.element)
         self.attr_id = self.get_attr_value('id')
         self.attr_class = self.get_attr_value('class').split()
-        self.bot.log.debug(self.CB_PROP_LOADED)
+        self.bot.log.debug(MSG.CB_PROP_LOADED)
         return True
+
+    def __check_reload__(self):
+        """Allow to check before methods calls to ensure
+            if it's neccessary reload element properties
+        """
+        if not self.element and self.auto_reload:
+            self.reload(**self.settings)
 
     def find_child(self, selector, locator=By.CSS_SELECTOR):
         """Find child element using bot with default By.CSS_SELECTOR strategy
@@ -210,16 +183,15 @@ class ControlBase(object):
         Returns:
             ControlBase -- instanced base element using qacode library object
         """
-        self.bot.log.debug(self.CB_FINDCHILD_LOADING.format(selector))
-        if not self.element and self.auto_reload:
-            self.reload(**self.settings)
+        self.bot.log.debug(MSG.CB_FINDCHILD_LOADING.format(selector))
+        self.__check_reload__()
         settings = {"locator": locator, "selector": selector}
         ele = self.bot.navigation.find_element_child(
             self.element, selector, locator=locator)
         settings.update({"element": ele})
         ctl = ControlBase(self.bot, **settings)
         if ctl:
-            self.bot.log.debug(self.CB_FINDCHILD_LOADED)
+            self.bot.log.debug(MSG.CB_FINDCHILD_LOADED)
         return ctl
 
     def find_children(self, selector, locator=By.CSS_SELECTOR):
@@ -238,9 +210,8 @@ class ControlBase(object):
             list(ControlBase) -- instanced list of base element using
                 qacode library object
         """
-        self.bot.log.debug(self.CB_FINDCHILD_LOADING.format(selector))
-        if not self.element and self.auto_reload:
-            self.reload(**self.settings)
+        self.bot.log.debug(MSG.CB_FINDCHILD_LOADING.format(selector))
+        self.__check_reload__()
         settings = {"locator": locator, "selector": selector}
         elements = self.bot.navigation.find_element_children(
             self.element, selector, locator=locator)
@@ -249,17 +220,16 @@ class ControlBase(object):
             settings.update({"element": ele})
             ctls.append(ControlBase(self.bot, **settings))
         if bool(ctls):
-            self.bot.log.debug(self.CB_FINDCHILD_LOADED)
+            self.bot.log.debug(MSG.CB_FINDCHILD_LOADED)
         return ctls
 
     def get_tag(self):
         """Returns tag_name from Webelement"""
-        self.bot.log.debug(self.CB_GETTAG_LOADING)
-        if not self.element and self.auto_reload:
-            self.reload(**self.settings)
+        self.bot.log.debug(MSG.CB_GETTAG_LOADING)
+        self.__check_reload__()
         tag_name = self.bot.navigation.ele_tag(self.element)
         if tag_name:
-            self.bot.log.debug(self.CB_GETTAG_LOADED.format(tag_name))
+            self.bot.log.debug(MSG.CB_GETTAG_LOADED.format(tag_name))
         return tag_name
 
     def type_text(self, text, clear=False):
@@ -271,9 +241,8 @@ class ControlBase(object):
         Keyword Arguments:
             clear {bool} -- clear text element at enable key (default: {False})
         """
-        self.bot.log.debug(self.CB_TYPETEXT_LOADING.format(text))
-        if not self.element and self.auto_reload:
-            self.reload(**self.settings)
+        self.bot.log.debug(MSG.CB_TYPETEXT_LOADING.format(text))
+        self.__check_reload__()
         if clear:
             self.clear()
         self.bot.navigation.ele_write(self.element, text)
@@ -281,27 +250,25 @@ class ControlBase(object):
 
     def clear(self):
         """Clear input element text value"""
-        self.bot.log.debug(self.CB_CLEAR_LOADING)
-        if not self.element and self.auto_reload:
-            self.reload(**self.settings)
+        self.bot.log.debug(MSG.CB_CLEAR_LOADING)
+        self.__check_reload__()
         self.bot.navigation.ele_clear(self.element)
-        self.bot.log.debug(self.CB_CLEAR_LOADED)
+        self.bot.log.debug(MSG.CB_CLEAR_LOADED)
 
     def click(self, retry=True):
         """Click on element"""
-        self.bot.log.debug(self.CB_CLICK_LOADING)
-        if not self.element and self.auto_reload:
-            self.reload(**self.settings)
+        self.bot.log.debug(MSG.CB_CLICK_LOADING)
+        self.__check_reload__()
         try:
             self.bot.navigation.ele_click(element=self.element)
         except (ElementNotVisibleException, NoSuchElementException) as err:
             if retry:
-                self.bot.log.warning(self.CB_CLICK_RETRY)
+                self.bot.log.warning(MSG.CB_CLICK_RETRY)
                 self.reload(**self.settings)
                 self.bot.navigation.ele_click(element=self.element)
             else:
                 raise err
-        self.bot.log.debug(self.CB_CLICK_LOADED)
+        self.bot.log.debug(MSG.CB_CLICK_LOADED)
 
     def get_text(self, on_screen=True):
         """Get element content text.
@@ -317,19 +284,18 @@ class ControlBase(object):
         Returns:
             str -- Return element content text (innerText property)
         """
-        self.bot.log.debug(self.CB_GETTEXT_LOADING)
-        if not self.element and self.auto_reload:
-            self.reload(**self.settings)
+        self.bot.log.debug(MSG.CB_GETTEXT_LOADING)
+        self.__check_reload__()
         text = None
         try:
             text = self.bot.navigation.ele_text(
                 self.element, on_screen=on_screen)
         except CoreException as err:
-            self.bot.log.error(self.CB_GETTEXT_FAILED)
-            raise ControlException(err, message=err.message)
+            self.bot.log.error(MSG.CB_GETTEXT_FAILED)
+            raise ControlException(err=err)
         if text:
             self.text = text
-            self.bot.log.debug(self.CB_GETTEXT_LOADED.format(text))
+            self.bot.log.debug(MSG.CB_GETTEXT_LOADED.format(text))
         return text
 
     def get_attrs(self, attr_names):
@@ -343,9 +309,8 @@ class ControlBase(object):
         Returns:
             dict -- a dict list of {name, value}
         """
-        self.bot.log.debug(self.CB_GETATTRS_LOADING)
-        if not self.element and self.auto_reload:
-            self.reload(**self.settings)
+        self.bot.log.debug(MSG.CB_GETATTRS_LOADING)
+        self.__check_reload__()
         attrs = list()
         for attr_name in attr_names:
             attrs.append({
@@ -365,18 +330,17 @@ class ControlBase(object):
         Returns:
             str -- value of html attr_name
         """
-        self.bot.log.debug(self.CB_GETATTRVALUE_LOADING.format(attr_name))
-        if not self.element and self.auto_reload:
-            self.reload(**self.settings)
+        self.bot.log.debug(MSG.CB_GETATTRVALUE_LOADING.format(attr_name))
+        self.__check_reload__()
         try:
             value = self.bot.navigation.ele_attribute(
                 self.element, attr_name)
             self.bot.log.debug(
-                self.CB_GETATTRVALUE_LOADED.format(attr_name, value))
+                MSG.CB_GETATTRVALUE_LOADED.format(attr_name, value))
             return value
         except CoreException as err:
-            self.bot.log.error(self.CB_GETATTRVALUE_FAILED)
-            raise ControlException(err, message=err.message)
+            self.bot.log.error(MSG.CB_GETATTRVALUE_FAILED)
+            raise ControlException(err=err)
 
     def get_css_value(self, prop_name):
         """Allows to obtain CSS value based on CSS property name
@@ -387,11 +351,10 @@ class ControlBase(object):
         Returns:
             str -- Value of CSS property searched
         """
-        self.bot.log.debug(self.CB_GETCSSRULE_LOADING)
-        if not self.element and self.auto_reload:
-            self.reload(**self.settings)
+        self.bot.log.debug(MSG.CB_GETCSSRULE_LOADING)
+        self.__check_reload__()
         css_value = self.bot.navigation.ele_css(self.element, prop_name)
-        self.bot.log.debug(self.CB_GETCSSRULE_LOADED.format(css_value))
+        self.bot.log.debug(MSG.CB_GETCSSRULE_LOADED.format(css_value))
         return css_value
 
     def set_css_value(self, prop_name, prop_value, css_important=True):
@@ -407,15 +370,14 @@ class ControlBase(object):
                 overrite others values applied (default: {True})
         """
         self.bot.log.debug(
-            self.CB_SETCSSRULE_LOADING.format(prop_name, prop_value))
-        if not self.element and self.auto_reload:
-            self.reload(**self.settings)
+            MSG.CB_SETCSSRULE_LOADING.format(prop_name, prop_value))
+        self.__check_reload__()
         self.bot.navigation.set_css_rule(
             self.selector, prop_name, prop_value,
             css_important=css_important)
         if self.selector is None:
-            self.bot.log.error(self.CB_SETCSSRULE_FAILED)
-            raise ControlException(message="Couldn't reload element")
+            self.bot.log.error(MSG.CB_SETCSSRULE_FAILED)
+            raise ControlException(msg="Couldn't reload element")
         self.reload(**self.settings)
 
     def reload(self, **kwargs):
@@ -423,7 +385,7 @@ class ControlBase(object):
             logic with new configuration
         """
         class_name = self.__class__.__name__
-        self.bot.log.debug(self.CB_RELOAD_LOADING.format(class_name))
+        self.bot.log.debug(MSG.CB_RELOAD_LOADING.format(class_name))
         # load settings again
         if kwargs:
             config = kwargs.copy()
@@ -438,7 +400,7 @@ class ControlBase(object):
             element=self.element)
         self._load_properties(enabled=self.on_instance_load)
         if class_name == 'ControlBase':
-            self.bot.log.debug(self.CB_RELOAD_LOADED.format(class_name))
+            self.bot.log.debug(MSG.CB_RELOAD_LOADED.format(class_name))
 
     def wait_invisible(self, timeout=0):
         """Wait for invisible element, returns control"""
@@ -448,8 +410,7 @@ class ControlBase(object):
 
     def wait_visible(self, timeout=0):
         """Wait for visible element, returns control"""
-        if not self.element and self.auto_reload:
-            self.reload(**self.settings)
+        self.__check_reload__()
         self.element = self.bot.navigation.ele_wait_visible(
             self.element, timeout=timeout)
         return self